restructure folder and update script commands
