+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "backend",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "server.js",
   "license": "MIT",
   "scripts": {
-    "start": "node build/server.js",
-    "dev": "npx nodemon -r dotenv/config server.ts dotenv_config_path=.env.local",
+    "dev": "ts-node-dev --exit-child src/server.ts",
+    "start": "ts-node src/server.ts",
     "copy-schemas": "$(yarn bin)/cpx **/*.graphql build",
     "build-ts": "tsc && yarn run copy-schemas",
     "watch-ts": "tsc -w",
@@ -21,27 +21,23 @@
     "express": "^4.17.1",
     "graphql": "^15.5.0",
     "graphql-middleware": "^6.0.7",
-    "merge-graphql-schemas": "^1.7.8"
+    "merge-graphql-schemas": "^1.7.8",
+    "ts-node-dev": "^1.1.8"
   },
   "devDependencies": {
-    "@babel/plugin-transform-runtime": "^7.13.15",
-    "@babel/preset-env": "^7.13.15",
-    "@babel/preset-typescript": "^7.13.0",
     "@types/express": "^4.17.11",
     "@types/graphql": "^14.5.0",
     "@types/jest": "^26.0.23",
     "@types/node": "^14.14.41",
     "@typescript-eslint/eslint-plugin": "^4.22.0",
     "@typescript-eslint/parser": "^4.22.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
     "cpx": "^1.5.0",
     "eslint": "^7.25.0",
     "eslint-plugin-import": "^2.22.1",
     "jest": "^26.6.3",
     "supertest": "^6.1.3",
     "ts-jest": "^26.5.5",
-    "ts-node": "^9.1.1",
+    "ts-node": "^10.7.0",
     "typescript": "^4.2.4"
   }
 }
