+++ src/@types/graphql.d.ts
