+++ src/graphql/schema/people.graphql
